Split drawing and moving into two actions
Drawing is separate from moving, and it makes it easier for me to think
about. This also lays the groundwork for doing connection processing on
the full set of devices, then afterwards drawing these connections.

The reason it didn't work before is that I created a counting while loop
without incrementing the counter.